refactor: 패키지 생성 및 코드 리팩터링
- signUp: 회원가입에 필요한 데이터 저장
- userClause: 이용약관 관련 데이터 처리 추가
- userPreference: 회원 프로필 수정 기능 구현 및 개선
