+++ app/src/main/java/kr/dori/android/own_cast/userClause/First3Fragment.kt
@@ -1,4 +1,4 @@
-package kr.dori.android.own_cast
+package kr.dori.android.own_cast.userClause
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,14 +6,15 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
-import kr.dori.android.own_cast.databinding.FragmentFirst2Binding
+import kr.dori.android.own_cast.R
+import kr.dori.android.own_cast.databinding.FragmentFirst3Binding
 
 /**
  * A simple [Fragment] subclass as the default destination in the navigation.
  */
-class First2Fragment : Fragment() {
+class First3Fragment : Fragment() {
 
-    private var _binding: FragmentFirst2Binding? = null
+    private var _binding: FragmentFirst3Binding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -24,7 +25,7 @@ class First2Fragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
 
-        _binding = FragmentFirst2Binding.inflate(inflater, container, false)
+        _binding = FragmentFirst3Binding.inflate(inflater, container, false)
         return binding.root
 
     }
@@ -33,7 +34,7 @@ class First2Fragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         binding.buttonFirst.setOnClickListener {
-            findNavController().navigate(R.id.action_First2Fragment_to_Second2Fragment)
+            findNavController().navigate(R.id.action_First3Fragment_to_Second3Fragment)
         }
     }
