+++ app/src/main/java/kr/dori/android/own_cast/userClause/Second2Fragment.kt
@@ -1,4 +1,4 @@
-package kr.dori.android.own_cast
+package kr.dori.android.own_cast.userClause
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,14 +6,15 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
-import kr.dori.android.own_cast.databinding.FragmentFirst3Binding
+import kr.dori.android.own_cast.R
+import kr.dori.android.own_cast.databinding.FragmentSecond2Binding
 
 /**
- * A simple [Fragment] subclass as the default destination in the navigation.
+ * A simple [Fragment] subclass as the second destination in the navigation.
  */
-class First3Fragment : Fragment() {
+class Second2Fragment : Fragment() {
 
-    private var _binding: FragmentFirst3Binding? = null
+    private var _binding: FragmentSecond2Binding? = null
 
     // This property is only valid between onCreateView and
     // onDestroyView.
@@ -24,7 +25,7 @@ class First3Fragment : Fragment() {
         savedInstanceState: Bundle?
     ): View {
 
-        _binding = FragmentFirst3Binding.inflate(inflater, container, false)
+        _binding = FragmentSecond2Binding.inflate(inflater, container, false)
         return binding.root
 
     }
@@ -32,8 +33,8 @@ class First3Fragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        binding.buttonFirst.setOnClickListener {
-            findNavController().navigate(R.id.action_First3Fragment_to_Second3Fragment)
+        binding.buttonSecond.setOnClickListener {
+            findNavController().navigate(R.id.action_Second2Fragment_to_First2Fragment)
         }
     }
