+++ app/src/main/java/kr/dori/android/own_cast/userClause/SecondFragment.kt
@@ -1,4 +1,4 @@
-package kr.dori.android.own_cast
+package kr.dori.android.own_cast.userClause
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
@@ -6,6 +6,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.navigation.fragment.findNavController
+import kr.dori.android.own_cast.R
 import kr.dori.android.own_cast.databinding.FragmentSecondBinding
 
 /**
