+++ app/src/main/java/kr/dori/android/own_cast/userPreference/ChangeDetailInterestActivity.kt
@@ -1,5 +1,4 @@
-package kr.dori.android.own_cast
-
+package kr.dori.android.own_cast.userPreference
 
 import android.content.Intent
 import android.os.Bundle
@@ -8,12 +7,30 @@ import android.view.View
 import android.widget.Button
 import android.widget.ImageView
 import android.widget.TextView
+import android.widget.Toast
 import androidx.activity.ComponentActivity
 import androidx.activity.enableEdgeToEdge
 import androidx.core.content.ContextCompat
+import retrofit2.Call
+import retrofit2.Callback
+import retrofit2.Response
 
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
+import kr.dori.android.own_cast.MainActivity
+import kr.dori.android.own_cast.PreferenceRequest
+import kr.dori.android.own_cast.PreferenceResponse
+import kr.dori.android.own_cast.R
+import kr.dori.android.own_cast.RetrofitClient
+import kr.dori.android.own_cast.SignupData
+import kr.dori.android.own_cast.forApiData.CastInterface
+import kr.dori.android.own_cast.forApiData.getRetrofit
+import kr.dori.android.own_cast.keyword.KeywordAppData
+import kr.dori.android.own_cast.keyword.KeywordLoadingDialog
 
-class SignupDetailInterestActivity : ComponentActivity() {
+class ChangeDetailInterestActivity : ComponentActivity() {
 
     // 각 체크박스 ImageView를 선언
     private lateinit var checkboxFirst: ImageView
@@ -31,10 +48,10 @@ class SignupDetailInterestActivity : ComponentActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_signup_detail_interest)
+        setContentView(R.layout.activity_change_detail_interest)
         enableEdgeToEdge()
 
-        val num = when (SignupData.interest) {
+        val num = when (SignupData.temp_interest) {
             "movie" -> 1
             "sports" -> 2
             "music" -> 3
@@ -104,22 +121,130 @@ class SignupDetailInterestActivity : ComponentActivity() {
         // 초기에는 완료 버튼을 비활성화
         confirmButton.isEnabled = false
         confirmButton.setOnClickListener {
-            startActivity(Intent(this, SignupLoadingActivity::class.java))
+
+            // interest 값을 변환
+            val mainCategory = when (SignupData.temp_interest) {
+                "movie" -> "드라마/영화"
+                "sports" -> "스포츠"
+                "music" -> "음악"
+                "food" -> "음식"
+                "book" -> "책"
+                "news" -> "시사/뉴스"
+                "art" -> "미술"
+                "self" -> "직접 입력"
+                else -> "드라마/영화"
+            }
+
+            // 요청 객체 생성
+            val preferenceRequest = PreferenceRequest(
+                mainCategory = mainCategory,
+                subCategory = SignupData.temp_detail_interest
+            )
+
+
+            // Retrofit 클라이언트 인스턴스를 통해 ApiService 호출
+            val apiService = RetrofitClient.instance
+
+            // 사용자 토큰 가져오기
+            val userToken = "Bearer ${SignupData.token}"
+
+            // 네트워크 호출, 토큰을 헤더에 포함
+            apiService.updatePreferences(userToken, preferenceRequest)
+                .enqueue(object : Callback<PreferenceResponse> {
+                    override fun onResponse(
+                        call: Call<PreferenceResponse>,
+                        response: Response<PreferenceResponse>
+                    ) {
+                        // 응답이 성공적이고 요청이 성공했는지 확인
+                        if (response.isSuccessful && response.body()?.isSuccess == true) {
+
+                            // 성공적인 응답 처리
+                            SignupData.profile_detail_interest = "완료"
+
+                            SignupData.interest = SignupData.temp_interest
+                            SignupData.detail_interest = SignupData.temp_detail_interest
+
+                            Toast.makeText(this@ChangeDetailInterestActivity, "관심사 변경 완료", Toast.LENGTH_SHORT).show()
+
+                            initData()
+
+
+
+                        } else {
+                            // 응답 오류 처리
+                            Toast.makeText(
+                                this@ChangeDetailInterestActivity,
+                                "서버 오류: ${response.body()?.message}",
+                                Toast.LENGTH_SHORT
+                            ).show()
+                        }
+                    }
+
+                    override fun onFailure(call: Call<PreferenceResponse>, t: Throwable) {
+                        // 네트워크 호출 실패 처리
+                        t.printStackTrace()
+                        Toast.makeText(
+                            this@ChangeDetailInterestActivity,
+                            "관심사 변경 실패",
+                            Toast.LENGTH_SHORT
+                        ).show()
+                    }
+                })
+
         }
+
 
         // 닫기 버튼
         findViewById<ImageView>(R.id.pop_exit_button).setOnClickListener {
-            val intent = Intent(this, SignupFifthActivity::class.java)
+            val intent = Intent(this, ChangeInterestActivity ::class.java)
             startActivity(intent)
+            finish()
         }
 
         // 직접 입력
         findViewById<TextView>(R.id.self).setOnClickListener {
-            val intent = Intent(this, SignupSelfActivity::class.java)
+            val intent = Intent(this, ChangeSelfActivity::class.java)
             startActivity(intent)
+            finish()
         }
     }
+    fun initData(){
 
+        Log.d("initDataCheck","${KeywordAppData.detailTopic.isNullOrEmpty()}, ${KeywordAppData.detailTopic}")
+
+        val getKeyword = getRetrofit().create(CastInterface::class.java)
+        val dialog = KeywordLoadingDialog(this,"데이터를 불러오고 있어요")
+        dialog.setCancelable(false)
+        dialog.setCanceledOnTouchOutside(false)
+        dialog.show()
+        CoroutineScope(Dispatchers.IO).launch() {
+            val response = getKeyword.getKeywordHome()
+            launch {
+                withContext(Dispatchers.Main) {
+                    try {
+                        dialog.dismiss()
+                        if (response.isSuccessful) {
+                            response.body()?.result?.let{
+                                KeywordAppData.updateDetailTopic(it.keywords)
+                            }
+                        } else {
+                            Log.d("initDataFinish","failed code : ${response.code()}")
+                            Log.d("initDataFinish","${KeywordAppData.detailTopic.size}")
+                            Toast.makeText(this@ChangeDetailInterestActivity,"데이터를 받아오는데 실패했습니다.\n 에러코드 : ${response.code()}",Toast.LENGTH_SHORT).show()
+                        }
+
+                    } catch (e: Exception) {
+                        e.printStackTrace()
+                    }
+                    val intent = Intent(this@ChangeDetailInterestActivity, MainActivity ::class.java)
+                    startActivity(intent)
+                    finish()
+                }
+
+
+            }
+        }
+    }
     private fun setSectionText(textViewId: Int, text: String) {
         val textView = findViewById<TextView>(textViewId)
         textView.text = text
@@ -129,7 +254,6 @@ class SignupDetailInterestActivity : ComponentActivity() {
     private fun onCheckboxClicked(view: View) {
         val clickedCheckbox = view as ImageView
         val clickedCheckboxId = clickedCheckbox.id
-        val sectionNumber = SignupData.interest
 
         // 클릭된 체크박스가 이미 선택된 상태인지 확인
         if (selectedCheckboxId == clickedCheckbox.id) {
@@ -137,14 +261,18 @@ class SignupDetailInterestActivity : ComponentActivity() {
             clickedCheckbox.setImageResource(R.drawable.signup_detail_uncheckbtn)
             selectedCheckboxId = null
             confirmButton.isEnabled = false
-            confirmButton.setBackgroundTintList(ContextCompat.getColorStateList(this, R.color.button_unclick))
+            confirmButton.setBackgroundTintList(ContextCompat.getColorStateList(this,
+                R.color.button_unclick
+            ))
         } else {
             // 새로운 체크박스를 선택하면 기존 선택을 해제하고 새로운 체크박스를 선택
             resetAllCheckboxes()
             clickedCheckbox.setImageResource(R.drawable.signup_detail_checkbtn)
             selectedCheckboxId = clickedCheckbox.id
             confirmButton.isEnabled = true
-            confirmButton.setBackgroundTintList(ContextCompat.getColorStateList(this, R.color.main_purple))
+            confirmButton.setBackgroundTintList(ContextCompat.getColorStateList(this,
+                R.color.main_purple
+            ))
 
             // 선택된 체크박스의 텍스트를 가져와 저장하기
             val checkboxIndex = when (clickedCheckboxId) {
@@ -178,13 +306,13 @@ class SignupDetailInterestActivity : ComponentActivity() {
                 )
 
                 // 현재 관심사 섹션 가져오기
-                val sectionInterests = interestsMap[SignupData.interest] ?: return
+                val sectionInterests = interestsMap[SignupData.temp_interest] ?: return
                 val index = checkboxIndex - 1
                 if (index in sectionInterests.indices) {
-                    SignupData.detail_interest = sectionInterests[index]
+                    SignupData.temp_detail_interest = sectionInterests[index]
                 } else {
                     // 인덱스가 범위를 벗어난 경우 기본값 설정
-                    SignupData.detail_interest = sectionInterests.last()
+                    SignupData.temp_detail_interest = sectionInterests.last()
                 }
             }
 
@@ -201,5 +329,4 @@ class SignupDetailInterestActivity : ComponentActivity() {
         checkboxFifth.setImageResource(R.drawable.signup_detail_uncheckbtn)
         checkboxSixth.setImageResource(R.drawable.signup_detail_uncheckbtn)
     }
-
 }
