+++ app/src/main/res/layout/acivity_change_password.xml
@@ -5,7 +5,7 @@
     android:id="@+id/profile_setting"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".PasswordSettingActivity">
+    tools:context=".signUp.PasswordSettingActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
