+++ app/src/main/res/layout/activity_change_interest.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ChangeInterestActivity">
+    tools:context=".userPreference.ChangeInterestActivity">
 
     <ScrollView
         android:layout_width="match_parent"
