+++ app/src/main/res/layout/activity_change_password.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ChangePasswordActivity">
+    tools:context=".userPreference.ChangePasswordActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
