+++ app/src/main/res/layout/activity_change_profile.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ProfileSettingActivity">
+    tools:context=".userPreference.ProfileSettingActivity">
 
     <RelativeLayout
         android:layout_width="match_parent"
