+++ app/src/main/res/layout/activity_signup_detail_interest.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".SignupDetailInterestActivity"
+    tools:context=".signUp.SignupDetailInterestActivity"
     android:background="#4C000000">
