+++ app/src/main/res/layout/activity_signup_first.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".SignupFirstActivity">
+    tools:context=".signUp.SignupFirstActivity">
 
 
     <LinearLayout
