+++ app/src/main/res/layout/activity_signup_language.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".SignupLanguageActivity">
+    tools:context=".signUp.SignupLanguageActivity">
 
     <ScrollView
         android:layout_width="match_parent"
