+++ app/src/main/res/layout/activity_signup_loading.xml
@@ -3,7 +3,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".SignupLoadingActivity">
+    tools:context=".signUp.SignupLoadingActivity">
 
     <FrameLayout
         android:id="@+id/progress_section"
