+++ app/src/main/res/layout/activity_signup_self.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".SignupSelfActivity">
+    tools:context=".signUp.SignupSelfActivity">
 
 
     <LinearLayout
