+++ app/src/main/res/layout/activity_signup_third.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".SignupThirdActivity">
+    tools:context=".signUp.SignupThirdActivity">
 
     <ScrollView
         android:id="@+id/scroll_view"
