+++ app/src/main/res/navigation/nav_graph2.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/First2Fragment"
-        android:name="kr.dori.android.own_cast.First2Fragment"
+        android:name="kr.dori.android.own_cast.userClause.First2Fragment"
         android:label="@string/first_fragment_label"
         tools:layout="@layout/fragment_first2">
 
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/Second2Fragment"
-        android:name="kr.dori.android.own_cast.Second2Fragment"
+        android:name="kr.dori.android.own_cast.userClause.Second2Fragment"
         android:label="@string/second_fragment_label"
         tools:layout="@layout/fragment_second2">
